Fix table update script.

+++ db/account_has_logs.microtime.col.sql
@@ -1,2 +1,2 @@
-ALTER TABLE account_has_logs CHANGE time microtime INT(16) UNSIGNED NOT NULL;
+ALTER TABLE account_has_logs CHANGE time microtime BIGINT(16) UNSIGNED NOT NULL;
 UPDATE account_has_logs SET microtime = microtime * 1000000;
